requirements.txt
# Core libraries
numpy==1.26.4
sounddevice==0.5.1
python-dotenv==1.0.1
requests==2.32.3

# Email and logging (standard libs, shown here for completeness)
# smtplib and logging are built into Python, no need to install

# Optional (if you want plotting support later)
matplotlib==3.9.0
